[aggregation] Move method to a separate module. Allow for subgroup special cases.
